third_party/WebKit/LayoutTests/fast/events/wheel/wheel-fling-cancel-expected.txt

Tests that a mouse wheel event will cancel an in progress fling gesture.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS actualWheelEventsOccurred is 0
PASS successfullyParsed is true

TEST COMPLETE
